backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
pillow
requests
tqdm
faiss-cpu
torch
transformers
python-multipart
sentencepiece
protobuf
